+++ Tests/ContextSpecs.swift
@@ -113,14 +113,14 @@ final class ContextSpecs: QuickSpec {
 
                 // ARRANGE
                 self.tested.changeState(state: self.mockState)
-                let media = ModernAVPlayerMedia(url: URL(string: "foo")!, type: .clip)
+                let media = PlayerMedia(url: URL(string: "foo")!, type: .clip)
 
                 // ACT
                 self.tested.loadMedia(media: media, autostart: false)
 
                 // ASSERT
                 expect(self.mockState.loadMedialCallCount).to(equal(1))
-                expect(self.mockState.lastMediaParam as? ModernAVPlayerMedia).to(equal(media))
+                expect(self.mockState.lastMediaParam).to(equal(media))
                 expect(self.mockState.lastShoudlPlayingParam).to(beFalse())
             }
         }